Refactored geom functions into PathGeomOp.
