+++ src/Mod/Path/TestPathApp.py
@@ -28,6 +28,7 @@ from PathTests.TestPathLog   import TestPathLog
 from PathTests.TestPathCore  import TestPathCore
 #from PathTests.TestPathPost  import PathPostTestCases
 from PathTests.TestPathGeom  import TestPathGeom
+from PathTests.TestPathGeomOp  import TestPathGeomOp
 from PathTests.TestPathUtil  import TestPathUtil
 from PathTests.TestPathDepthParams        import depthTestCases
 from PathTests.TestPathDressupHoldingTags import TestHoldingTags
